+++ .idea/misc.xml
@@ -8,9 +8,10 @@
         <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/favorites_fragment.xml" value="0.3625" />
         <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/favorites_fragment2.xml" value="0.3625" />
         <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/films_fragment.xml" value="0.3625" />
-        <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/fragment_main.xml" value="0.3625" />
-        <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/series_fragment.xml" value="0.3625" />
+        <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/fragment_main.xml" value="0.2" />
+        <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/series_fragment.xml" value="0.34524929444967073" />
         <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/show_item.xml" value="0.5" />
+        <entry key="..\:/Users/david/Documents/GitHub/android-challenge/app/src/main/res/layout/toolbar_main.xml" value="0.3625" />
       </map>
     </option>
   </component>
