+++ app/src/main/java/com/zygotecnologia/zygotv/ui/main/MainFragment.kt
@@ -1,28 +1,17 @@
 package com.zygotecnologia.zygotv.ui.main
 
 import android.os.Bundle
-import android.util.Log
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.bumptech.glide.Glide
-import com.bumptech.glide.request.RequestOptions
+import androidx.fragment.app.Fragment
+import com.google.android.material.tabs.TabLayoutMediator
 import com.zygotecnologia.zygotv.R
-import com.zygotecnologia.zygotv.data.enum.GenresIdEnum
-import com.zygotecnologia.zygotv.data.model.GenreResponse
-import com.zygotecnologia.zygotv.data.model.Show
-import com.zygotecnologia.zygotv.data.model.ShowResponse
 import com.zygotecnologia.zygotv.databinding.FragmentMainBinding
-import com.zygotecnologia.zygotv.ui.MainViewModel
-import com.zygotecnologia.zygotv.utils.ImageUrlBuilder
-import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class MainFragment : Fragment() {
 
     private lateinit var binding: FragmentMainBinding
-
-    private val viewModel: MainViewModel by viewModel()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -35,73 +24,28 @@ class MainFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-
-        initObservers()
-
-        loadShows()
+        configViewPager()
     }
 
-    private fun initObservers(){
-        viewModel.genres.observe(viewLifecycleOwner, ::onGetGenres)
-        viewModel.shows.observe(viewLifecycleOwner, ::onGetShows)
-    }
-
-    private fun onGetGenres(genreResponse: GenreResponse?){
-        if(genreResponse != null){
-            viewModel.getShows()
-        }
-    }
-
-    private fun onGetShows(showResponse: ShowResponse?){
-        val listShows = showResponse?.results?: emptyList()
-
-        configRecyclerViews(listShows)
-        configImageBanner(listShows)
-    }
-
-    private fun loadShows(){
-        viewModel.getGenres()
-    }
-
-    private fun configRecyclerViews(listShows: List<Show>){
-        binding.rvShowListComedy.apply {
-            this.adapter = MainAdapter(listShows.filter { show -> show.genreIds?.contains(
-                GenresIdEnum.COMEDY.generoId) ?: false })
-        }
-
-        binding.rvShowListAdventure.apply {
-            this.adapter = MainAdapter(listShows.filter { show -> show.genreIds?.contains(
-                GenresIdEnum.ACTION_ADVENTURE.generoId) ?: false })
-        }
-
-        binding.rvShowListCrime.apply {
-            this.adapter = MainAdapter(listShows.filter { show -> show.genreIds?.contains(
-                GenresIdEnum.CRIME.generoId) ?: false })
-        }
-
-    }
-    private fun configImageBanner(listShows: List<Show>){
-        var mostPopuplarVotes = 0
-        var mostPopularId: Int? = 0
-
-        listShows.forEach { show ->
-            show.voteCount?.let {
-                if (it > mostPopuplarVotes) {
-                    mostPopularId = show.id
-                    mostPopuplarVotes = it
-                }
+    private fun configTabLayout(){
+        TabLayoutMediator(binding.tabLayout, binding.viewPager){ tab, position ->
+            when(position){
+                0 -> tab.text = getString(R.string.tab_item_films)
+                1 -> tab.text = getString(R.string.tab_item_series)
+                else -> tab.text = getString(R.string.tab_item_favorites)
             }
-        }
 
-        val mostPopularShow = listShows.filter { it.id == mostPopularId }
-
-        listShows.map { Log.d("###", "$mostPopularShow") }
-
-        Glide.with(binding.ivShowPopular)
-            .load(mostPopularShow[0].backdropPath?.let { ImageUrlBuilder.buildBackdropUrl(it) })
-            .apply(RequestOptions().placeholder(R.drawable.image_placeholder))
-            .into(binding.ivShowPopular)
-
-        binding.tvTitleShowPopular.text = mostPopularShow[0].name
+        }.attach()
     }
+
+    private fun configViewPager(){
+        //desabilitando o swipe do viewPager, e tornando a navegação possível somente pelas tabs
+        binding.viewPager.also {
+            it.isUserInputEnabled = false
+            it.adapter = MainActivityAdapter(requireActivity())
+            it.currentItem = 1
+        }
+        configTabLayout()
+    }
+
 }
