+++ app/src/main/res/values-night/themes.xml
@@ -2,15 +2,20 @@
     <!-- Base application theme. -->
     <style name="Theme.ZygoTV" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
         <!-- Primary brand color. -->
-        <item name="colorPrimary">@color/purple_200</item>
-        <item name="colorPrimaryVariant">@color/purple_700</item>
-        <item name="colorOnPrimary">@color/black</item>
-        <!-- Secondary brand color. -->
-        <item name="colorSecondary">@color/teal_200</item>
-        <item name="colorSecondaryVariant">@color/teal_200</item>
-        <item name="colorOnSecondary">@color/black</item>
+        <item name="colorPrimary">@color/gray_primary</item>
+        <item name="colorPrimaryDark">@color/gray_primary</item>
+        <item name="colorAccent">@color/white</item>
         <!-- Status bar color. -->
         <item name="android:statusBarColor" tools:targetApi="l">@color/black</item>
         <!-- Customize your theme here. -->
+
+        <item name="android:colorBackground">@color/gray_primary</item>
+        <item name="tabStyle">@style/Theme.ZygoTV.TabLayout</item>
+    </style>
+
+    <style name="Theme.ZygoTV.TabLayout" parent="Widget.Design.TabLayout">
+        <item name="tabBackground">@color/gray_primary</item>
+        <item name="tabSelectedTextColor">@color/white</item>
+        <item name="tabTextColor">@color/gray_secondary</item>
     </style>
 </resources>
