+++ app/src/main/res/values/colors.xml
@@ -8,4 +8,5 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="gray_primary">#151515</color>
+    <color name="gray_secondary">#9A9A9A</color>
 </resources>
