+++ app/src/main/res/values/dimens.xml
@@ -6,8 +6,10 @@
 
     <dimen name="text_size_16">16sp</dimen>
 
-    <dimen name="item_list_poster_path_width">140dp</dimen>
-    <dimen name="item_list_poster_path_height">200dp</dimen>
+    <dimen name="item_list_poster_path_width">120dp</dimen>
+    <dimen name="item_list_poster_path_height">190dp</dimen>
 
     <dimen name="corner_radius_default">7dp</dimen>
+
+    <dimen name="banner_height">170dp</dimen>
 </resources>
