+++ app/src/main/res/values/strings.xml
@@ -8,4 +8,8 @@
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
 
+    <!-- Tabs -->
+    <string name="tab_item_films">Films</string>
+    <string name="tab_item_series">Series</string>
+    <string name="tab_item_favorites">Favorites</string>
 </resources>
