+++ app/src/main/res/values/themes.xml
@@ -1,18 +1,21 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.ZygoTV" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+    <style name="Theme.ZygoTV" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/gray_primary</item>
-        <item name="colorPrimaryVariant">@color/purple_700</item>
-        <item name="colorOnPrimary">@color/white</item>
-        <!-- Secondary brand color. -->
-        <item name="colorSecondary">@color/teal_200</item>
-        <item name="colorSecondaryVariant">@color/teal_700</item>
-        <item name="colorOnSecondary">@color/black</item>
+        <item name="colorPrimaryDark">@color/gray_primary</item>
+        <item name="colorAccent">@color/white</item>
         <!-- Status bar color. -->
         <item name="android:statusBarColor" tools:targetApi="l">@color/black</item>
         <!-- Customize your theme here. -->
 
         <item name="android:colorBackground">@color/gray_primary</item>
+        <item name="tabStyle">@style/Theme.ZygoTV.TabLayout</item>
+    </style>
+
+    <style name="Theme.ZygoTV.TabLayout" parent="Widget.Design.TabLayout">
+        <item name="tabBackground">@color/gray_primary</item>
+        <item name="tabSelectedTextColor">@color/white</item>
+        <item name="tabTextColor">@color/gray_secondary</item>
     </style>
 </resources>
